medicare_assistant/requirements.txt
langchain>=0.2.0
langgraph>=0.2.0
langchain-groq>=0.1.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
streamlit>=1.35.0
ragas>=0.1.0
datasets>=2.19.0
python-dotenv>=1.0.0
eval-type-backport>=0.2.0
jupyter>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
